feat: refactor ResourceTags component
